requirements.txt
rasterio
google-generativeai
geopandas
shapely
Pillow
tqdm
python-dotenv
